Revert "moved commands.h file to network folder"
This reverts commit e94affeadf9053bb2445ef68e15d0c31f6b2012b.
